+++ docs/changelog.md
@@ -4,7 +4,8 @@
 
 **Features**:
 
-+ Fix joint merging function.
++ Fix joint merging function. (#43)
++ Fix infinity loop of solver. (#44)
 + Enhance "Remove empty links" function. ("Remove redundant links")
 
 **Development**: