+++ docs/changelog.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Version 21.01
+
+**Features**:
+
++ None.
+
+**Development**:
+
++ None.
+
 ## Version 20.12
 
 **Features**: